fixed spk_wvf.get_waveform last float32

+++ npyx/spk_wvf.py
@@ -218,7 +218,7 @@ def get_waveforms(dp, u, n_waveforms=100, t_waveforms=82, subset_selection='regu
     # Correct voltage scaling
     waveforms*=read_spikeglx_meta(dp, 'ap')['scale_factor']
         
-    return  waveforms.astype(np.float16)
+    return  waveforms.astype(np.float32)
 
 def get_w(traces, slc, _n_samples_extract):
     # Get slice
